Move save palette to projects container
