+++ src/containers/Projects/Projects.js
@@ -1,11 +1,12 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
-import { fetchOptionsCreator } from '../../utils/fetchOptionsCreator';
-import { postNewProject } from '../../thunks/postNewProject';
-import closeButton from '../../assets/closeButton.png';
 import { storePalettes, storeProjects, setMessage } from '../../actions';
+import closeButton from '../../assets/closeButton.png';
 import { deletePalette } from '../../thunks/deletePalette';
 import { deleteProject } from '../../thunks/deleteProject';
+import { postNewPalette } from '../../thunks/postNewPalette';
+import { postNewProject } from '../../thunks/postNewProject';
+import { fetchOptionsCreator } from '../../utils/fetchOptionsCreator';
 import PropTypes from 'prop-types';
 
 export class Projects extends Component {
@@ -13,11 +14,17 @@ export class Projects extends Component {
     super();
     this.state = {
       project_name: '',
-      project_id: 0
+      project_id: 0,
+      palette_name: '',
+      color_1: '',
+      color_2: '',
+      color_3: '',
+      color_4: '',
+      color_5: ''
     }
   }
 
-  updateProjectName = (e) => {
+  handleChange = (e) => {
     const { name, value } = e.target
     this.setState({ [name]: value })
   }
@@ -50,7 +57,35 @@ export class Projects extends Component {
     this.props.deleteProject(id)
   }
 
+  setPalette = async (e) => {
+    e.preventDefault()
+    await this.props.currentColors.forEach((color, index) => {
+      let colorKey = `color_${index + 1}`
+      this.setState({ [colorKey]: color.color })
+    })
+    this.addNewPalette()
+  }
+  
+  addNewPalette = async () => {
+    const { project_id, palette_name, color_1, color_2, color_3, color_4, color_5 } = this.state
+    const body = {
+      project_id,
+      palette_name,
+      color_1,
+      color_2,
+      color_3,
+      color_4,
+      color_5
+    }
+    const options = await fetchOptionsCreator('POST', body)
+    this.props.postNewPalette(options, body)
+  }
+
   render() {
+    const projectList = this.props.projects.map((project, index) => {
+      let projectName = project.project_name
+      return <option key={`option-${projectName}-${index}`} value={project.project_id} >{projectName}</option>
+    })
     const projectsToDisplay = this.props.projects.map((project, index) => {
       return <div className="individual-projects" key={`${project.project_name}-${index}`}>
         <h3 className="project-title">{project.project_name}</h3>
@@ -83,13 +118,21 @@ export class Projects extends Component {
     return(
       <div className="projects">
       <form onSubmit={this.addNewProject}>
-        <input type="text" onChange={this.updateProjectName} value={this.state.project_name} name="project_name" />
+        <input type="text" onChange={this.handleChange} value={this.state.project_name} name="project_name" />
         <button className="project-button">Add New Project</button>
         <div className="message-container">
         {this.props.message.length > 0 ? (
           <p className="user-message">{this.props.message}</p>) : (<p></p>
         )}
         </div>
+      </form>
+      <select value={this.state.project_id} name="project_id" onChange={this.handleChange}>
+        <option value="0" disable="true" select="true" default>Select a Project</option>
+        {projectList}
+      </select>
+      <form onSubmit={this.setPalette}>
+        <input type="text" onChange={this.handleChange} value={this.state.palette_name} name="palette_name" />
+        <button className="palette-button" disabled={this.state.project_id == 0}>Save Palette to Project</button>
       </form>
       <div className="projects-to-display">
         <p className="my-projects-title">my projects</p>
@@ -110,11 +153,14 @@ Projects.propTypes = {
   storeProjects: PropTypes.func.isRequired,
   deleteProject: PropTypes.func.isRequired,
   setMessage: PropTypes.func.isRequired,
+  currentColors: PropTypes.array.isRequired,
+  postNewPalette: PropTypes.func.isRequired
 }
 
 export const mapStateToProps = (state) => ({
   projects: state.projects,
   palettes: state.palettes,
+  currentColors: state.currentColors,
   message: state.message
 })
 
@@ -124,7 +170,8 @@ export const mapDispatchToProps = (dispatch) => ({
   deletePalette: (id) => dispatch(deletePalette(id)),
   storeProjects: (updatedProjects) => dispatch(storeProjects(updatedProjects)),
   deleteProject: (id) => dispatch(deleteProject(id)),
-  setMessage: (message) => dispatch(setMessage(message))
+  setMessage: (message) => dispatch(setMessage(message)),
+  postNewPalette: (options, body) => dispatch(postNewPalette(options, body))
 })
 
 export default connect(mapStateToProps, mapDispatchToProps)(Projects)
